# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pengine Local API",
    "version": "1.0.0",
    "description": "Local-only API for persona selection, observable asynchronous short-drama creation, timeout and final-review recovery, complete result retrieval, and one frozen revision."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Loopback-only local service"
    }
  ],
  "security": [],
  "paths": {
    "/personas": {
      "get": {
        "operationId": "listPersonas",
        "summary": "List valid selectable persona packages",
        "responses": {
          "200": {
            "description": "Selectable personas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaList"
                },
                "examples": {
                  "nonProduction": {
                    "value": {
                      "items": [
                        {
                          "persona_id": "example-non-production",
                          "display_name": "非生产示例人格",
                          "version": "0.0.0-example",
                          "snapshot_sha256": "e1bf5a772e1a10851a9abd801323966c339204802d55918fb0877d48a690fcde"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "pengine_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pengine Session"
            }
          }
        ]
      }
    },
    "/creations": {
      "post": {
        "operationId": "createCreation",
        "summary": "Queue an initial complete creation run",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "pengine_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pengine Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCreationRequest"
              },
              "examples": {
                "nonProduction": {
                  "value": {
                    "persona_id": "example-non-production",
                    "story": "一个离开故乡多年的人回乡处理旧屋。",
                    "requirements": "创作一部当代短剧；具体集数和单集标准从所选人格的 L4 获取。"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Initial run queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreationAccepted"
                },
                "examples": {
                  "queued": {
                    "value": {
                      "creation_id": "018f6d49-2e10-7b21-8f40-5f5162b9d181",
                      "initial_state": "queued",
                      "resource_url": "/creations/018f6d49-2e10-7b21-8f40-5f5162b9d181"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/PersonaNotFound"
          },
          "409": {
            "$ref": "#/components/responses/IdempotencyConflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Creations",
        "operationId": "listCreations",
        "parameters": [
          {
            "name": "pengine_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pengine Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreationList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandError"
                }
              }
            }
          }
        }
      }
    },
    "/creations/{creation_id}": {
      "get": {
        "operationId": "getCreation",
        "summary": "Get initial and revision states and complete results",
        "parameters": [
          {
            "$ref": "#/components/parameters/CreationId"
          },
          {
            "name": "pengine_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pengine Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Creation resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreationResource"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/CreationNotFound"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteCreation",
        "summary": "Delete Creation",
        "parameters": [
          {
            "$ref": "#/components/parameters/CreationId"
          },
          {
            "name": "pengine_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pengine Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "$ref": "#/components/responses/CreationNotFound"
          },
          "409": {
            "description": "The creation still has queued or running work",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandError"
                },
                "example": {
                  "code": "creation_not_deletable",
                  "message": "The creation still has queued or running work."
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandError"
                }
              }
            }
          }
        }
      }
    },
    "/creations/{creation_id}/revision": {
      "post": {
        "operationId": "createOrRetryRevision",
        "summary": "Freeze and queue the only revision, or requeue its identical feedback after failure",
        "description": "The first accepted feedback is frozen. A failed revision may be requeued only with byte-for-byte identical feedback. Queued, running, or succeeded revisions reject this command.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CreationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "pengine_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pengine Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequest"
              },
              "examples": {
                "nonProduction": {
                  "value": {
                    "feedback": "主角最后的选择太顺利，请让代价更明确，同时保留原有情绪底色。"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Revision run queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionAccepted"
                },
                "examples": {
                  "queued": {
                    "value": {
                      "creation_id": "018f6d49-2e10-7b21-8f40-5f5162b9d181",
                      "revision_state": "queued",
                      "feedback_locked": true,
                      "resource_url": "/creations/018f6d49-2e10-7b21-8f40-5f5162b9d181"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/CreationNotFound"
          },
          "409": {
            "description": "Idempotency conflict, revision unavailable, or feedback differs from the frozen feedback",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandError"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandError"
                }
              }
            }
          }
        }
      }
    },
    "/creations/{creation_id}/runs/{run_kind}/continue": {
      "post": {
        "summary": "Continue a paused initial or revision run",
        "operationId": "continueRun",
        "parameters": [
          {
            "name": "creation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Creation Id"
            }
          },
          {
            "name": "run_kind",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "initial",
                "revision"
              ],
              "type": "string",
              "title": "Run Kind"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "pengine_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pengine Session"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Run queued or an identical continue command replayed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunControlAccepted"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/CreationNotFound"
          },
          "409": {
            "description": "The requested run cannot continue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandError"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "A caller-generated key. Reuse with the same request replays the original command response; reuse with a different request returns idempotency_conflict.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      }
    },
    "responses": {
...
      "ServiceUnavailable": {
        "description": "Local service cannot accept the command",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CommandError"
            },
            "example": {
              "code": "service_unavailable",
              "message": "The service is temporarily unavailable."
            }
          }
        }
      }
    },
    "schemas": {
...
      "CreationAccepted": {
        "properties": {
          "creation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Creation Id"
          },
          "initial_state": {
            "type": "string",
            "const": "queued",
            "title": "Initial State",
            "default": "queued"
          },
          "resource_url": {
            "type": "string",
            "title": "Resource Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "creation_id",
          "resource_url"
        ],
        "title": "CreationAccepted"
      },
...
      "PersonaList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PersonaSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PersonaList"
      },
...
    }
  },
...
}